+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix Guide page layout [#116](https://github.com/hushenghao/AndroidEasterEggs/issues/116)
 - Migrate kapt to KSP
 - Fix Android ICS egg abnormal finish
+- Fix F-Droid Reproducible Builds
 
 ### v2.2.1 (2023-11-25)
 - Fix F-Droid Reproducible Builds